+++ src/probe/bmp.rs
@@ -1,23 +1,21 @@
 //! Black Magic Probe interface.
 
+use super::{
+    begin_monitor_output, gdb_script_command, gdb_script_continue, gdb_script_wait, run_gdb_client,
+    rustc_substitute_path, setup_uart_endpoint,
+};
 use crate::{
     cli::{ProbeFlashCmd, ProbeGdbCmd, ProbeMonitorCmd, ProbeResetCmd},
-    itm,
-    probe::{run_gdb_client, rustc_substitute_path, setup_uart_endpoint},
+    monitor,
     templates::Registry,
     utils::{block_with_signals, exhaust_fifo, make_fifo, run_command, spawn_command, temp_dir},
 };
 use anyhow::{anyhow, Result};
 use drone_config as config;
 use signal_hook::iterator::Signals;
-use std::{
-    fs::OpenOptions,
-    io::{Read, Write},
-    path::{Path, PathBuf},
-    process::Command,
-};
+use std::{fs::File, path::PathBuf, thread};
 use tempfile::tempdir_in;
-use termcolor::{Color, ColorSpec, StandardStream, WriteColor};
+use termcolor::StandardStream;
 
 /// Black Magic Probe `drone probe reset` command.
 #[allow(missing_docs)]
@@ -35,11 +33,7 @@ impl ResetCmd<'_> {
         let Self { cmd, signals, registry, config, config_probe } = self;
         let ProbeResetCmd {} = cmd;
         let script = registry.bmp_reset(config)?;
-        let mut gdb = Command::new(&config_probe.gdb_client);
-        gdb.arg("--quiet");
-        gdb.arg("--nx");
-        gdb.arg("--batch");
-        gdb.arg("--command").arg(script.path());
+        let gdb = gdb_script_command(config_probe, None, script.path());
         block_with_signals(&signals, true, || run_command(gdb))
     }
 }
@@ -60,12 +54,7 @@ impl FlashCmd<'_> {
         let Self { cmd, signals, registry, config, config_probe } = self;
         let ProbeFlashCmd { firmware } = cmd;
         let script = registry.bmp_flash(config)?;
-        let mut gdb = Command::new(&config_probe.gdb_client);
-        gdb.arg(firmware);
-        gdb.arg("--quiet");
-        gdb.arg("--nx");
-        gdb.arg("--batch");
-        gdb.arg("--command").arg(script.path());
+        let gdb = gdb_script_command(config_probe, Some(firmware), script.path());
         block_with_signals(&signals, true, || run_command(gdb))
     }
 }
@@ -97,9 +86,9 @@ impl GdbCmd<'_> {
     }
 }
 
-/// Black Magic Probe `drone probe monitor` command.
+/// Black Magic Probe `drone probe monitor` SWO command.
 #[allow(missing_docs)]
-pub struct MonitorCmd<'a> {
+pub struct MonitorSwoCmd<'a> {
     pub cmd: &'a ProbeMonitorCmd,
     pub signals: Signals,
     pub registry: Registry<'a>,
@@ -109,7 +98,7 @@ pub struct MonitorCmd<'a> {
     pub shell: &'a mut StandardStream,
 }
 
-impl MonitorCmd<'_> {
+impl MonitorSwoCmd<'_> {
     /// Runs the command.
     pub fn run(self) -> Result<()> {
         let Self { cmd, signals, registry, config, config_probe, config_probe_swo, shell } = self;
@@ -127,32 +116,18 @@ impl MonitorCmd<'_> {
         let pipe = make_fifo(&dir)?;
         let ports = outputs.iter().flat_map(|output| output.ports.iter().copied()).collect();
         let script = registry.bmp_swo(config, &ports, *reset, &pipe)?;
-        let mut gdb = Command::new(&config_probe.gdb_client);
-        gdb.arg("--quiet");
-        gdb.arg("--nx");
-        gdb.arg("--batch");
-        gdb.arg("--command").arg(script.path());
-        let mut gdb = spawn_command(gdb)?;
-
-        let (pipe, packet) = block_with_signals(&signals, false, move || {
-            let mut packet = [0];
-            OpenOptions::new().read(true).open(&pipe)?.read_exact(&mut packet)?;
-            Ok((pipe, packet))
-        })?;
+        let mut gdb = spawn_command(gdb_script_command(config_probe, None, script.path()))?;
+        let (pipe, packet) = gdb_script_wait(&signals, pipe)?;
 
         exhaust_fifo(uart_endpoint)?;
-        itm::spawn(&Path::new(uart_endpoint), outputs);
+        let input = File::open(uart_endpoint)?;
+        let outputs = monitor::Output::open_all(outputs)?;
+        thread::spawn(move || {
+            monitor::swo::capture(input, &outputs);
+        });
 
-        block_with_signals(&signals, false, move || {
-            OpenOptions::new().write(true).open(&pipe)?.write_all(&packet)?;
-            Ok(())
-        })?;
-
-        shell.set_color(ColorSpec::new().set_bold(true).set_fg(Some(Color::Cyan)))?;
-        writeln!(shell)?;
-        writeln!(shell, "{:=^80}", " ITM OUTPUT ")?;
-        shell.reset()?;
-
+        gdb_script_continue(&signals, pipe, packet)?;
+        begin_monitor_output(shell)?;
         block_with_signals(&signals, true, move || {
             gdb.wait()?;
             Ok(())